feat: make cloth config optional
