+++ fabric/build.gradle
@@ -20,6 +20,7 @@ configurations {
 }
 
 dependencies {
+    implementation 'org.testng:testng:7.1.0'
     modImplementation "net.fabricmc:fabric-loader:${rootProject.fabric_loader_version}"
     modImplementation "net.fabricmc.fabric-api:fabric-api:${rootProject.fabric_api_version}"
