+++ fabric/src/main/java/net/nimajnebec/autoelytra/fabric/ModMenuIntegration.java
@@ -2,7 +2,7 @@ package net.nimajnebec.autoelytra.fabric;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
-import net.nimajnebec.autoelytra.Configuration;
+import net.nimajnebec.autoelytra.config.Configuration;
 
 public class ModMenuIntegration implements ModMenuApi {
     @Override
