+++ fabric/src/main/resources/fabric.mod.json
@@ -25,5 +25,9 @@
     "fabricloader": ">=0.14.21",
     "minecraft": ">=1.20.1",
     "fabric": "*"
+  },
+
+  "suggests": {
+    "cloth-config": "*"
   }
 }
