+++ forge/src/main/java/net/nimajnebec/autoelytra/config/forge/ConfigurationImpl.java
@@ -1,0 +1,7 @@
+package net.nimajnebec.autoelytra.config.forge;
+
+public class ConfigurationImpl {
+    public static boolean ClothConfigInstalled() {
+        return false;
+    }
+}
